added stream tests to hw 5

+++ 05/long.cpp
@@ -6,7 +6,7 @@
 
 using namespace std;
 
-Long::Long(int a)
+Long::Long(long long a)
 {
     if (a == 0)
     {
